counterpart area supported
Function wb_data_flow allows you to specify counterpart area(s). This is useful for variables of the IDS that do not work with the original wb_data function because they require a counterpart area and are accessed with the "series" url. I added wb_counterparts and wb_series to the cache. Limited capacity with queries at the moment.